🎨 Feat:"data upload api 구조 개선"
Resolves: #17
