+++ data/crawling_divorce_all_0708 - crawling_divorce_cleaned_v8.csv
@@ -1,0 +1,8 @@
+판례일련번호,사건명,사건번호,법원명,판결요지,판례내용
+116830,이혼·이혼및위자료,98므15,대법원,"  혼인생활의 파탄에 대하여 주된 책임이 있는 배우자는 원칙적으로 그 파탄을 사유로 하여 이혼을 청구할 수 없고, 다만 상대방도 그 파탄 이후 혼인을 계속할 의사가 없음이 객관적으로 명백한데도 오기나 보복적 감정에서 이혼에 응하지 아니하고 있을 뿐이라는 등 특별한 사정이 있는 경우에만 예외적으로 이혼을 청구할 수 있으며, 유책배우자의 이혼청구에 대하여 상대방이 그 주장사실을 다투면서 오히려 다른 사실을 내세워 반소로 이혼청구를 한다 하더라도 그러한 사정만으로 곧바로 상대방은 혼인을 계속할 의사가 없으면서도 오기나 보복적 감정에서 유책배우자의 이혼청구에 응하지 아니하는 것이라고 단정할 수 없다. ","원고(반소피고),상고인 피고(반소원고),피상고인 원심판결 
+서울고법  1997. 11. 13. 선고 97르1242, 1259 판결주문  상고를 기각한다.  상고비용은 원고(반소피고)의 부담으로 한다.이유  1. 원심은, 원고(반소피고, 다음부터는 원고라 한다)와 피고(반소원고, 다음부터는 피고라 한다) 1951. 3. 13. 고향에서 중매로 혼인하여 2남 3녀를 두었으나 서로 성격이 맞지 아니하여 가정불화가 심화되던 중 원고는 피고가 무능하고 지능이 낮다는 이유로 걸핏하면 피고에게 욕설을 하고 피고의 머리채를 잡아 끌면서 집을 나가라고 협박하고 피고의 시아버지도 피고에게 집을 나가라고 하여 피고는 더 이상 참지 못하고 1962.경 집을 나와 친정 근처에 집을 얻어 살면서 현재까지 35년 동안 혼자 어렵게 생활하고 있는 사실, 원고는 장남과 장녀를 고향에 남겨두고 나머지 자식만 데리고 서울로 와서 
+        소외인과 동거생활을 하며 그 사이에 새로 1남 2녀를 두었고, 사업을 하여 상당한 재산도 모은 사실 등을 인정한 다음, 원·피고 사이의 혼인생활이 파탄에 이른 주된 책임은 원고에게 있음을 이유로 이혼을 구하는 원고의 이 사건 본소청구를 배척하고, 피고의 반소청구에 따라 원·피고는 이혼하고 원고는 피고에게 위자료로 금 150,000,000원을 지급할 것을 명하였다.
+        2. 가. 기록과 대조하여 살펴보면, 원·피고 사이의 혼인이 파탄에 이르게 된 경위와 그 원인에 관한 원심의 사실인정과 판단은 정당하고, 거기에 채증법칙을 위배한 사실오인의 위법 등이 없다. 따라서 이 점에 관한 상고이유는 받아들이지 아니한다.  나. 
+        혼인생활의 파탄에 대하여 주된 책임이 있는 배우자는 원칙적으로 그 파탄을 사유로 하여 이혼을 청구할 수 없고, 다만 상대방도 그 파탄 이후 혼인을 계속할 의사가 없음이 객관적으로 명백한데도 오기나 보복적 감정에서 이혼에 응하지 아니하고 있을 뿐이라는 등 특별한 사정이 있는 경우에만 예외적으로 이혼을 청구할 수 있으며(대법원 1997. 5. 16. 선고 97므155 판결 참조), 
+        유책배우자의 이혼청구에 대하여 상대방이 그 주장사실을 다투면서 오히려 다른 사실을 내세워 반소로 이혼청구를 한다 하더라도 그러한 사정만으로 곧바로 상대방은 혼인을 계속할 의사가 없으면서도 오기나 보복적 감정에서 유책배우자의 이혼청구에 응하지 아니하는 것이라고 단정할 수 없다.
+        이 사건에서 보면, 피고는 자녀들이 이혼에 반대하고 또 아무런 생계수단이 없는 상태에서 이혼에 응할 수 없다는 이유로 원고의 이혼요구를 거부하여 왔음을 알 수 있고, 또 피고는 지금까지 자식들에게 누가 될 것 같아 원고의 이혼요구를 거부하였으나 원고가 이 사건 본소를 제기한 후 자식들에게 말로 다하지 못하는 행패를 부리는 것을 보고 부모로서 더 이상 부모 자식간에 의가 상하는 것을 두고 볼 수 없어 이혼을 결심하기에 이르렀다고 주장하고 있는 점 등에 비추어 보면, 피고는 원고와의 혼인을 계속할 의사가 없으면서도 오기나 보복적 감정에서 원고의 이혼청구에 응하지 아니한 것이라고 볼 수 없다.  원심이 같은 취지에서, 피고의 이 사건 반소청구에도 불구하고 혼인생활의 파탄에 주된 책임이 있는 원고는 스스로 이혼청구를 할 수 없다고 하여 이 사건 본소청구를 배척한 조치는 옳고, 거기에 유책배우자의 이혼청구권에 관한 법리 등을 오해한 위법이 없다. 원고가 내세우는 이 법원의 판결들은 이 사건과는 사안을 달리하는 것이어서 이 사건 해결에 도움이 되지 아니한다. 따라서 이 점에 관한 상고이유도 받아들이지 아니한다.  다. 원심이 인정한 사실관계와 그 밖에 이 사건 변론에 나타난 모든 사정에 비추어 보면, 원심이 인정한 위자료의 액수는 수긍이 가고, 거기에 위자료 산정에 관한 법리오해의 위법이 없다. 따라서 이 점에 관한 상고이유도 받아들이지 아니한다.  3. 그러므로 상고를 기각하고 상고비용은 패소자의 부담으로 하여 주문과 같이 판결한다.대법관 이돈희(재판장) 최종영 이임수 서성(주심) "
