+++ data/crawling_divorce_all_0708 - crawling_divorce_cleaned_v9.csv
@@ -1,0 +1,17 @@
+판례일련번호,사건명,사건번호,법원명,판결요지,판례내용
+146713,이혼및위자료등·이혼등,97므1486,대법원,"  [1] 민법 제839조의2에 규정된 재산분할제도는 혼인 중에 취득한 실질적인 공동재산을 청산 분배하는 것을 주된 목적으로 하는 것이므로, 부부가 재판상 이혼을 할 때 쌍방의 협력으로 이룩한 재산이 있는 한, 법원으로서는 당사자의 청구에 의하여 그 재산의 형성에 기여한 정도 등 당사자 쌍방의 일체의 사정을 참작하여 분할의 액수와 방법을 정하여야 하는바, 이 경우 부부 일방의 특유재산은 원칙적으로 분할의 대상이 되지 아니하나 특유재산일지라도 다른 일방이 적극적으로 그 특유재산의 유지에 협력하여 그 감소를 방지하였거나 그 증식에 협력하였다고 인정되는 경우에는 분할의 대상이 될 수 있다.  [2] 부부 일방이 혼인 중 제3자에게 부담한 채무는 일상가사에 관한 것 이외에는 원칙으로 그 개인의 채무로서 청산의 대상이 되지 않으나 그것이 공동재산의 형성에 수반하여 부담한 채무인 경우에는 청산의 대상이 된다.  [3] 재산분할의 방법이나 그 비율 또는 액수는 당사자 쌍방의 협력으로 이룩한 재산의 액수 기타의 사정을 참작하여 법원이 이를 정하되 그 모든 사정을 개별적·구체적으로 일일이 특정하여 설시하여야 하는 것은 아니다.","원고,피상고인(부대상고인,반소피고) 피고,상고인(부대피상고인,반소원고) 사건본인 원심판결 
+서울고법  1997. 9. 12. 선고 97르447, 454 판결주문상고 및 부대상고를 각 기각한다.  상고비용은 피고(반소원고)의, 부대상고비용은 원고(반소피고)의 각 부담으로 한다.이유  상고 및 부대상고이유를 함께 본다.민법 제839조의2에 규정된 재산분할제도는 혼인 중에 취득한 실질적인 공동재산을 청산 분배하는 것을 주된 목적으로 하는 것이므로, 부부가 재판상 이혼을 할 때 쌍방의 협력으로 이룩한 재산이 있는 한, 법원으로서는 당사자의 청구에 의하여 그 재산의 형성에 기여한 정도 등 당사자 쌍방의 일체의 사정을 참작하여 분할의 액수와 방법을 정하여야 하는바{대법원 1995. 10. 12. 선고 95므175(본소), 95므182(반소) 판결 등 참조}, 
+
+        이 경우 부부 일방의 특유재산은 원칙적으로 분할의 대상이 되지 아니하나 특유재산일지라도 다른 일방이 적극적으로 그 특유재산의 유지에 협력하여 그 감소를 방지하였거나 그 증식에 협력하였다고 인정되는 경우에는 분할의 대상이 될 수 있고, 한편 부부 일방이 혼인 중 제3자에게 부담한 채무는 일상가사에 관한 것 이외에는 원칙으로 그 개인의 채무로서 청산의 대상이 되지 않으나 그것이 공동재산의 형성에 수반하여 부담한 채무인 경우에는 청산의 대상이 된다고 할 것이며, 
+
+        재산분할의 방법이나 그 비율 또는 액수는 당사자 쌍방의 협력으로 이룩한 재산의 액수 기타의 사정을 참작하여 법원이 이를 정하되 그 모든 사정을 개별적·구체적으로 일일이 특정하여 설시하여야 하는 것은 아니다(대법원 1993. 5. 25. 선고 92므501 판결 등 참조).
+
+        원심판결 이유에 의하면, 원심은 거시 증거에 의하여, 피고(반소원고, 이하 피고라 한다)가 고등학교를 졸업한 후 곧바로 취업하여 얻은 수입으로 결혼 전에 원심 판시 단독주택을 매입함에 있어서 피고의 모(母)인 
+
+        제1심 공동피고 명의로 소유권이전등기를 마쳤다가 이를 매각하여 얻은 처분대금과 은행융자금을 주된 자원으로 하여 결혼한 지 5년만에 현재 위 
+
+        제1심 공동피고와 함께 거주하고 있는 원심 판시 쌍용아파트를 매수한 사실을 인정한 다음, 위 
+
+        제1심 공동피고 명의로 등기된 위 쌍용아파트를 재산분할의 대상으로 삼았는바, 기록에 의하면, 원심의 위와 같은 사실인정은 수긍이 가고, 사실관계가 원심이 인정한 바와 같다면 원고(반소피고, 이하 원고라 한다)가 가사노동 등을 통하여 직·간접으로 재산의 유지 및 증가에 기여하였다고 보이는 이상 위 쌍용아파트는 부부 쌍방의 협력으로 이룩한 재산으로 재산분할의 대상이 된다고 본 원심판단은 정당하고, 소론과 같은 위법이 있다고 할 수 없다. 
+
+        또한 원심이 분할대상에 포함시킨 경북 성주군 초전면 보정동에 있는 2필지 토지는 피고가 조모로부터 상속받은 특유재산임은 소론과 같으나, 기록에 의하면, 원고는 주로 가사노동에 종사하였지만 피고가 1989. 12. 16. 뇌졸증으로 쓰러져 1990. 5.경 그 동안 근무하여 오던 직장에서 퇴직한 후에는 보석학원에 다닌 경험을 살려 피고와 함께 금은방을 경영하기도 하였고 1994. 7.경부터는 도배공으로 일하면서 그 수입을 생활비로 사용하기도 한 사실을 인정할 수 있어 위 특유재산의 감소방지에 일정한 기여를 하였다고 할 수 있으므로, 위 특유재산을 재산분할의 대상으로 삼은 원심판단도 역시 수긍이 가고, 한편 원심이 거시 증거에 의하여 피고가 1992. 10.부터 원고가 가출한 1996. 5. 26.까지 피고의 사촌누이로부터 매달 금 1,000,000원씩 생활비를 차용한 사실을 인정하고 그로 인한 채무를 소극재산으로 적극재산에서 공제한 반면, 피고가 증권투자를 하면서 그 자금마련을 위하여 산업은행 및 상업은행 공덕동지점으로부터 차용하였다고 주장한 합계 금 34,200,000원의 채무는 이를 공제대상에서 제외시키면서 그 내용을 구체적으로 특정하여 설시하지 아니하였는바, 위에서 본 법리와 기록에 의하면 원심의 이러한 조치는 수긍이 가고, 거기에 소론과 같은 채증법칙 위배로 인한 사실오인이나 재산분할에 관한 법리오인의 위법이 있다고 할 수 없다. 논지는 모두 이유 없다.  그러므로 상고 및 부대상고를 각 기각하고 상고 및 부대상고비용은 각 패소자의 부담으로 하기로 하여 관여 법관의 일치된 의견으로 주문과 같이 판결한다.대법관 이용훈(재판장) 정귀호 박준서(주심) 김형선 "
